requirements-server-openi.txt
numpy
pandas
Pillow
PyYAML
scikit-learn
tqdm
fuzzywuzzy
python-Levenshtein
transformers
